Embedding tp (#19)
* Changes for all reduce Megatron style

* Changes for all reduce Megatron style

* Use TP Embedding component

* Use TP Embedding component

+++ metaseq/modules/transformer_layer.py
@@ -498,7 +498,7 @@ class TransformerDecoderLayer(nn.Module):
         """
         if need_head_weights:
             need_attn = True
-
+        x = x.half()
         residual = x
 
         if self.normalize_before:
